refactor: pass symbol-functables around instead of symbol-names.
In zz, there's no need for symbol-names in general, we can
just pass the Lua function itself around.  That lends itself
to a nice refactor to minimise use of symbol-names internally
in zmacs and zz, instead passing the symbol-functables.
* lib/zz/default-bindings.lua: Change second argument from
symbol-name to Lua function name.
* lib/zz/commands/bind.zl (global_set_key): Change second
parameter to accept a Lua function and bind it in the keymap.
* lib/zz/callbacks.lua: Likewise.
* lib/zmacs/eval.lua (get_function_doc): Remove unused function.
* lib/zz/eval.lua (function_exists, get_function_by_name)
(get_function_doc): Remove unused functions.

Signed-off-by: Gary V. Vaughan <gary@gnu.org>