zile: fix assert in switch_to_buffer, for after set-buffer.
If a ZLisp program changes the buffer_stack with `set-buffer` or
otherwise, and `switch_to_buffer` is called before reaching the
command read loop, handle the stack correctly.
* lib/zile/buffer.lua (switch_to_buffer): Don't fail the initial
assert if the current window still holds the buffer on the bottom
of the stack.
If the buffer for the current window is successfully changed,
then switch the buffer on the bottom of the stack too so that
when the stack unwinds, the buffer doesn't change back again.

Signed-off-by: Gary V. Vaughan <gary@gnu.org>