astr: use libc memrchr if available.
Trickier than it should be, because alien throws an error as
soon as Lua reads `alien.default.memrchr` before giving us a
chance to catch it and fallback to the brute force array search.
* lib/zile.astr.lua (memrchr): Use pcall and loadstring to safely
evaluate the alien interface to libc `memrchr`, but silently fall
back to the brute force reverse search if that fails.

Signed-off-by: Gary V. Vaughan <gary@gnu.org>